binary search running

+++ jac/jaclang/runtimelib/jacpim_simulation_runtime/dpu_support.h
@@ -13,7 +13,7 @@ void save(void * buf, uint32_t start, uint32_t size) {
 
 void run_thread(uint64_t walker_container_ptr, uint64_t trace_length, char * node_buffer, char * walker_buffer) {
     ContainerObject container_obj;
-    for (uint64_t i = 0; i < trace_length; i++) {
+    for (uint32_t i = 0; i < trace_length; i++) {
         get(&container_obj, walker_container_ptr + i * sizeof(ContainerObject), sizeof(ContainerObject));
         #ifdef DEBUG
         printf("DPU Tasklet %u: Container Object - Walker ptr: %lu, Walker size: %lu, Node ptr: %lu, Node size: %lu, Edge num: %lu, Func call: %lu\n", me(), container_obj.walker_ptr, container_obj.walker_size, container_obj.node_ptr, container_obj.node_size, container_obj.edge_num, container_obj.func_call);
